+++ src/Symfony/Bridge/Twig/Extension/CsrfExtension.php
@@ -17,6 +17,8 @@ use Twig\TwigFunction;
 /**
  * @author Christian Flothmann <christian.flothmann@sensiolabs.de>
  * @author Titouan Galopin <galopintitouan@gmail.com>
+ *
+ * @final since Symfony 4.4
  */
 class CsrfExtension extends AbstractExtension
 {