+++ src/Symfony/Bridge/Twig/TokenParser/TransDefaultDomainTokenParser.php
@@ -20,6 +20,8 @@ use Twig\TokenParser\AbstractTokenParser;
  * Token Parser for the 'trans_default_domain' tag.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @final since Symfony 4.4
  */
 class TransDefaultDomainTokenParser extends AbstractTokenParser
 {