+++ src/Symfony/Component/Mime/Tests/AddressTest.php
@@ -13,7 +13,6 @@ namespace Symfony\Component\Mime\Tests;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Mime\Address;
-use Symfony\Component\Mime\NamedAddress;
 
 class AddressTest extends TestCase
 {
@@ -22,6 +21,12 @@ class AddressTest extends TestCase
         $a = new Address('fabien@symfonï.com');
         $this->assertEquals('fabien@symfonï.com', $a->getAddress());
         $this->assertEquals('fabien@xn--symfon-nwa.com', $a->toString());
+        $this->assertEquals('fabien@xn--symfon-nwa.com', $a->getEncodedAddress());
+
+        $a = new Address('fabien@symfonï.com', 'Fabien');
+        $this->assertEquals('Fabien', $a->getName());
+        $this->assertEquals('fabien@symfonï.com', $a->getAddress());
+        $this->assertEquals('Fabien <fabien@xn--symfon-nwa.com>', $a->toString());
         $this->assertEquals('fabien@xn--symfon-nwa.com', $a->getEncodedAddress());
     }
 
@@ -34,7 +39,7 @@ class AddressTest extends TestCase
     public function testCreate()
     {
         $this->assertSame($a = new Address('fabien@symfony.com'), Address::create($a));
-        $this->assertSame($b = new NamedAddress('helene@symfony.com', 'Helene'), Address::create($b));
+        $this->assertSame($b = new Address('helene@symfony.com', 'Helene'), Address::create($b));
         $this->assertEquals($a, Address::create('fabien@symfony.com'));
     }
 
@@ -47,7 +52,7 @@ class AddressTest extends TestCase
     public function testCreateArray()
     {
         $fabien = new Address('fabien@symfony.com');
-        $helene = new NamedAddress('helene@symfony.com', 'Helene');
+        $helene = new Address('helene@symfony.com', 'Helene');
         $this->assertSame([$fabien, $helene], Address::createArray([$fabien, $helene]));
 
         $this->assertEquals([$fabien], Address::createArray(['fabien@symfony.com']));
@@ -57,5 +62,19 @@ class AddressTest extends TestCase
     {
         $this->expectException(\InvalidArgumentException::class);
         Address::createArray([new \stdClass()]);
+    }
+
+    /**
+     * @dataProvider nameEmptyDataProvider
+     */
+    public function testNameEmpty(string $name)
+    {
+        $mail = 'mail@example.org';
+        $this->assertSame($mail, (new Address($mail, $name))->toString());
+    }
+
+    public function nameEmptyDataProvider(): array
+    {
+        return [[''], [' '], [" \r\n "]];
     }
 }