+++ src/Symfony/Component/Translation/Tests/TranslatorTest.php
@@ -34,9 +34,12 @@ class TranslatorTest extends TestCase
     {
         $translator = new Translator($locale);
 
-        $this->assertEquals($locale, $translator->getLocale());
+        $this->assertSame($locale, $translator->getLocale());
     }
 
+    /**
+     * @group legacy
+     */
     public function testConstructorWithoutLocale()
     {
         $translator = new Translator(null);
@@ -73,6 +76,17 @@ class TranslatorTest extends TestCase
         $translator->setLocale($locale);
 
         $this->assertEquals($locale, $translator->getLocale());
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testSetNullLocale()
+    {
+        $translator = new Translator('en');
+        $translator->setLocale(null);
+
+        $this->assertNull($translator->getLocale());
     }
 
     public function testGetCatalogue()
@@ -158,6 +172,17 @@ class TranslatorTest extends TestCase
         $this->addToAssertionCount(1);
     }
 
+    /**
+     * @group legacy
+     */
+    public function testSetNullFallbackLocale()
+    {
+        $translator = new Translator('en');
+        $translator->setFallbackLocales(['fr', null]);
+        // no assertion. this method just asserts that no exception is thrown
+        $this->addToAssertionCount(1);
+    }
+
     public function testTransWithFallbackLocale()
     {
         $translator = new Translator('fr_FR');
@@ -184,9 +209,6 @@ class TranslatorTest extends TestCase
      */
     public function testAddResourceValidLocales($locale)
     {
-        if (null === $locale) {
-            $this->markTestSkipped('null is not a valid locale');
-        }
         $translator = new Translator('fr');
         $translator->addResource('array', ['foo' => 'foofoo'], $locale);
         // no assertion. this method just asserts that no exception is thrown
@@ -372,9 +394,6 @@ class TranslatorTest extends TestCase
      */
     public function testTransValidLocale($locale)
     {
-        if (null === $locale) {
-            $this->markTestSkipped('null is not a valid locale');
-        }
         $translator = new Translator($locale);
         $translator->addLoader('array', new ArrayLoader());
         $translator->addResource('array', ['test' => 'OK'], $locale);
@@ -393,6 +412,62 @@ class TranslatorTest extends TestCase
         $translator->addResource('array', $messages, 'fr', '');
 
         $this->assertEquals($expected, $translator->trans($id, [], '', 'fr'));
+    }
+
+    /**
+     * @dataProvider getTransChoiceTests
+     * @group legacy
+     */
+    public function testTransChoice($expected, $id, $translation, $number, $parameters, $locale, $domain)
+    {
+        $translator = new Translator('en');
+        $translator->addLoader('array', new ArrayLoader());
+        $translator->addResource('array', [(string) $id => $translation], $locale, $domain);
+
+        $this->assertEquals($expected, $translator->transChoice($id, $number, $parameters, $domain, $locale));
+    }
+
+    /**
+     * @dataProvider getInvalidLocalesTests
+     * @group legacy
+     */
+    public function testTransChoiceInvalidLocale($locale)
+    {
+        $this->expectException('Symfony\Component\Translation\Exception\InvalidArgumentException');
+        $translator = new Translator('en');
+        $translator->addLoader('array', new ArrayLoader());
+        $translator->addResource('array', ['foo' => 'foofoo'], 'en');
+
+        $translator->transChoice('foo', 1, [], '', $locale);
+    }
+
+    /**
+     * @dataProvider getValidLocalesTests
+     * @group legacy
+     */
+    public function testTransChoiceValidLocale($locale)
+    {
+        $translator = new Translator('en');
+        $translator->addLoader('array', new ArrayLoader());
+        $translator->addResource('array', ['foo' => 'foofoo'], 'en');
+
+        $translator->transChoice('foo', 1, [], '', $locale);
+        // no assertion. this method just asserts that no exception is thrown
+        $this->addToAssertionCount(1);
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testTransChoiceNullLocale()
+    {
+        $translator = new Translator('en');
+        $translator->addLoader('array', new ArrayLoader());
+        $translator->addResource('array', ['foo' => 'foofoo'], 'en');
+
+        $translator->transChoice('foo', 1, [], '', null);
+        // no assertion. this method just asserts that no exception is thrown
+        $this->addToAssertionCount(1);
     }
 
     public function getTransFileTests()
@@ -463,7 +538,6 @@ class TranslatorTest extends TestCase
     {
         return [
             [''],
-            [null],
             ['fr'],
             ['francais'],
             ['FR'],