+++ src/Symfony/Component/Mime/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 4.4.0
 -----
 
+ * [BC BREAK] Removed `NamedAddress` (`Address` now supports a name)
  * Added PHPUnit constraints
  * Added `AbstractPart::asDebugString()`